fix: code review improvements
