+++ src/javascript/ContentEditor/editorTabs/EditPanelContent/SidePanel/ContentHistory/ContentHistory.scss
@@ -24,8 +24,8 @@
     display: flex;
     flex-direction: row;
     gap: var(--spacing-small);
-    padding: var(--spacing-small);
     align-items: center;
+    padding: var(--spacing-small);
 
     border-bottom: 1px solid var(--color-gray_light);
