+++ src/javascript/ContentEditor/editorTabs/EditPanelContent/SidePanel/registerSidePanelTabs.js
@@ -1,7 +1,6 @@
 import React from 'react';
 import {sidePanelTabAction} from './sidePanelTabAction';
 import {InfoCircle, History, Visibility} from '@jahia/moonstone';
-import {Constants} from '~/ContentEditor/ContentEditor.constants';
 import {ContentDetails} from './ContentDetails';
 import {ContentHistory} from './ContentHistory';
 import {Preview} from '../Preview';
@@ -12,17 +11,16 @@ export const registerSidePanelTabs = actionsRegistry => {
         buttonLabel: 'jcontent:label.contentEditor.sidePanel.details',
         buttonIcon: <InfoCircle/>,
         targets: ['sidePanelTabsActions:2'],
-        value: 'details',
         dataSelRole: 'tab-details',
         displayableComponent: ContentDetails,
         isDisplayable: () => true,
+        requiredPermission: ['canSeeAdvancedOptionsTab']
     });
 
     actionsRegistry.add('action', 'ceSidePanelHistoryTab', sidePanelTabAction, {
         buttonLabel: 'jcontent:label.contentEditor.sidePanel.history',
         buttonIcon: <History/>,
         targets: ['sidePanelTabsActions:3'],
-        value: 'history',
         dataSelRole: 'tab-history',
         displayableComponent: ContentHistory,
         isDisplayable: () => true,
@@ -33,9 +31,8 @@ export const registerSidePanelTabs = actionsRegistry => {
         buttonLabel: 'jcontent:label.contentEditor.sidePanel.preview',
         buttonIcon: <Visibility/>,
         targets: ['sidePanelTabsActions:1'],
-        value: 'preview',
         dataSelRole: 'tab-preview',
         displayableComponent: Preview,
-        isDisplayable: ({hasPreview}) => Boolean(hasPreview),
+        isDisplayable: ({hasPreview}) => Boolean(hasPreview)
     });
 };
